# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nourish API
  description: >
    AI-powered nutrition & food tracking API. Every route except /health and
    /api/auth/register|login requires a Bearer JWT and is scoped to the
    authenticated user. Import this file directly into Postman, Insomnia, or
    Swagger UI (swagger.io/tools/swagger-ui) to explore and call it.
  version: "1.0.0"
servers:
  - url: http://localhost:4000/api
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: invalid_request }
            message: { type: string }
            details: { type: object, nullable: true }
...
    MealItem:
      type: object
      required: [name, qty, unit, source]
      properties:
        foodId: { type: string, format: uuid, nullable: true, description: "If set, macros are recomputed server-side and any macro fields you send are ignored." }
        name: { type: string }
        qty: { type: number }
        unit: { type: string, example: g }
        calories: { type: number, description: "Required only when foodId is absent." }
        proteinG: { type: number }
        carbsG: { type: number }
        fatG: { type: number }
        fibreG: { type: number }
        source: { type: string, enum: [manual, ai_image, voice, ocr, saved, ai_suggestion] }
        confidence: { type: number, nullable: true, minimum: 0, maximum: 1 }
    Meal:
      type: object
      properties:
        id: { type: string, format: uuid }
        date: { type: string, format: date }
        time: { type: string, example: "08:30" }
        mealType: { type: string, enum: [Breakfast, "Morning snack", Lunch, "Evening snack", Dinner, Custom] }
        source: { type: string }
        note: { type: string, nullable: true }
        imageAssetUrl: { type: string, nullable: true }
        items:
          type: array
          items: { $ref: "#/components/schemas/MealItem" }
  responses:
    Unauthorized:
      description: Missing/invalid token
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Not found (or not owned by the caller — existence is never leaked)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Invalid:
      description: Failed validation
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
paths:
...
  /auth/register:
    post:
      summary: Create an account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
      responses:
        "201": { description: "Created — returns { token, user }" }
        "400": { $ref: "#/components/responses/Invalid" }
        "409": { description: Email already in use }
  /auth/login:
    post:
      summary: Log in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200": { description: "OK — returns { token, user }" }
        "401": { description: Invalid credentials }
  /auth/me:
    get:
      summary: Current user
      responses:
        "200": { description: OK }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      summary: Permanently delete the account and all owned data (cascades via the DB schema)
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /foods/search:
    get:
      summary: Search the food catalog (reference DB + your custom foods)
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /foods/{id}:
    get:
      summary: Get one food with nutrients + servings
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "404": { $ref: "#/components/responses/NotFound" }
  /foods/custom:
    get:
      summary: List every custom food you own (manual entries + saved label-scan products)
      responses: { "200": { description: OK } }
    post:
      summary: Create a custom food (manual entry or a saved OCR product)
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        "201": { description: Created }
  /meals:
    get:
      summary: List meals (by date, or start/end range, optional text search)
      parameters:
        - { name: date, in: query, schema: { type: string, format: date } }
        - { name: start, in: query, schema: { type: string, format: date } }
        - { name: end, in: query, schema: { type: string, format: date } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  meals: { type: array, items: { $ref: "#/components/schemas/Meal" } }
    post:
      summary: Log a meal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, mealType, source, items]
              properties:
                date: { type: string, format: date }
                time: { type: string }
                mealType: { type: string }
                source: { type: string }
                note: { type: string }
                imageAssetUrl: { type: string, nullable: true }
                items: { type: array, items: { $ref: "#/components/schemas/MealItem" } }
      responses:
        "201": { description: Created }
        "400": { $ref: "#/components/responses/Invalid" }
  /meals/{id}:
    get:
      summary: Get one meal
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      summary: Update a meal (partial)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses:
        "200": { description: OK }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Delete a meal (soft delete)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /meals/{id}/duplicate:
    post:
      summary: Duplicate a meal, optionally to a different date
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { date: { type: string, format: date } } } } } }
      responses:
        "201": { description: Created }
...
  /favorites:
    get:
      summary: List favourites (foods, meals, recipes, products)
      responses: { "200": { description: OK } }
    post:
      summary: Add a favourite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, refId, name]
              properties:
                type: { type: string, enum: [food, meal, recipe, product] }
                refId: { type: string }
                name: { type: string }
                snapshot: { type: object }
      responses: { "201": { description: Created } }
  /favorites/{id}:
    delete:
      summary: Remove a favourite
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /analytics/range:
    get:
      summary: Aggregate stats for a date range (averages, consistency, top foods, weekday/weekend split)
      parameters:
        - { name: start, in: query, required: true, schema: { type: string, format: date } }
        - { name: end, in: query, required: true, schema: { type: string, format: date } }
      responses: { "200": { description: OK } }
  /analytics/day:
    get:
      summary: Totals + meals for a single date
      parameters: [{ name: date, in: query, required: true, schema: { type: string, format: date } }]
      responses: { "200": { description: OK } }
...